split LandingPage components and adjust empty list handling
