+++ src/App.js
@@ -1,5 +1,5 @@
 import { React } from "react";
-import Landing from "./components/Landing";
+import Landing from "./components/LandingPage/Landing";
 import Header from "./components/Layout/Header";
 
 function App() {
